changed variable names and paths

+++ L1_Soc_PPI.sh
@@ -10,21 +10,21 @@ MAINDATADIR=`pwd`/data
 MAINOUTPUTDIR=`pwd`/outputs
 cd $basedir
 
-#datadir=/home/tue90350/data/186_Subjects_${task}/$subj/${subj}_3T_tfMRI_${task}_preproc/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}
-
 OUTPUT=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_Social_PPI
 DATA=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/smoothing.feat/ICA_AROMA/denoised_func_data_nonaggr.nii.gz
 NVOLUMES=`fslnvols ${DATA}`
-rm -rf ${OUTPUT}.feat
+
+#remove if output file exists
+if [ -e ${OUTPUT}.feat ]; then
+  rm -rf ${OUTPUT}.feat
+fi
 
 #EV files
 EVMENTAL=${MAINDATADIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/EVs/mental.txt-
 EVRND=${MAINDATADIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/EVs/rnd.txt
 
-#Mask
-MASK=${basedir}/Masks/rT1_vPAC_Seed.nii
-
 #generate mask's timecourse
+MASK=${basedir}/Masks/rT1_vPAC_Seed.nii
 TIMECOURSE=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/my_timecourse.txt
 fslmeants -i $DATA -o $TIMECOURSE -m $MASK
 
@@ -34,12 +34,11 @@ OTEMPLATE=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_
 sed -e 's@OUTPUT@'$OUTPUT'@g' \
 -e 's@DATA@'$DATA'@g' \
 -e 's@NVOLUMES@'$NVOLUMES'@g' \
--e 's@EVFEAR@'$EVFEAR'@g' \
--e 's@EVNEUT@'$EVNEUT'@g' \
+-e 's@EVMENTAL@'$EVMENTAL'@g' \
+-e 's@EVRND@'$EVRND'@g' \
 -e 's@TIMECOURSE@'$TIMECOURSE'@g' \
 <$ITEMPLATE> $OTEMPLATE
-feat ${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_Social_PPI.fsf
-
+feat $OTEMPLATE
 
 #delete unused files
 #rm -rf ${OUTPUT}.feat/filtered_func_data.nii.gz
